Remove old saiten visualization. Create editor one And apply in editor note playfield.
